Added application for calculating perturbation vectors
The application RNApvmin provides an user interface to calculate
a perturbation vector for sequenced with additional SHAPE information.
The sequence is read from stdin while the SHAPE file has to be passed
as argument.
The user may influence the weighting of the objective function and the
gradient evaluation by setting the arguments sigma/tau and sampleSize.
The calculated perturbation vector will be written to stdout.
The current implementation relies on SHAPE data, that has alredy been
normalized to a pairing probability.

Signed-off-by: Ronny Lorenz <ronny@tbi.univie.ac.at>