Moved fprintf from vrna_pf_fold to RNAfold
vrna_pf_fold currently prints the ensemble energy to stderr when the
sequence has more than 1600 nucleotides.
This is annoying when processing large sequences with RNApvmin, which uses
stderr to print the progress of the minimization process. Since the
algorithm for finding an appropriate perturbation vector calls
vrna_pf_fold multiple times, the ensemble energy is printed very
often.
Moved the corresponding code chunk to RNAfold to solve this issue.