Removed redundant make_pscores from alifold.c and alipfold.c to data_structures.c
Since we create a ready-to-use data structure there, and we dont want to
make it a global API call, this is the best solution.
Also moved readribosum() from alifold.c to ribo.c where it belongs to.
Additionally, some consensus structure prediction related macro
constants moved to energy_const.

IMPORTANT: Hard constraints do not work for alifold anymore! Repair this in
constraints.c!