avoid overwriting of $seq1 in test.pl with newer swig version
Somehow, the change from swig 2 to 3 involves some call by reference
or whatever transition at least in the perl interfaces created. Therefore,
the assignment $struct = $seq1 and a successive call of RNA::fold($seq1,$struct) results in having the backtraced structure in $struct as well as $seq1.
This again makes all the following tests of the perl interface failing.

!It should be checked whether a reverse patch using swig2 really handles
this differently!