Made stochastic backtracking in part_func.c faster using Boustrophedon Scheme
The Boustrophedon scheme will not be applied to backtracing in the QB matrix,
since on real data we see a large increase in computation time.
On the other hand, large speedup is gained as well by precomputing the
Q1k, and Qln arrays, and reversing one or the other loop.