Enable LTO support for Linux with clang/clang++
The trick is to call clang/clang++ with -fuse-ld=gold for linking and
set the corresponding tool prefixes, i.e. llvm-ar, llvm-ranlib, and
llvm-nm.
This, however, required several changes in the autotools mess. We
now pass several variables to the subpackages to make them actually
compile and link to our static RNAlib with LTO support.

How this should be done for OS X remains unsolved for now.

This commit also sets RNAforester to a version that actually builds
with clang