Fixed bug in backtrack() in mfe.c that returned arbitrary data
This bug was only effecting the return value of the new
function vrna_backtrack_from_intervals() in mfe.c, since it is the only
one that uses, and passes along the return value of backtrack().
Nevertheless, this was a bug, so we fixed this by setting the return
type of both functions to void.