Add an ordered stream output data structure with unordered fill capability
This data structure, vrna_ostream_t and the associated functions allow
one to process data in an ordered fashion while still producing and
inserting the data into the stream in an unordered manner. This is
especially useful if parallel processing is in place but output needs
to be in sequential order. As soon as consecutive data blocks from the
beginning of the stream become available, they are sequentially passed
to the callback in the right order. The implementation currently uses
pthread_mutex_t to allow for threadsafe memory management and
callback processing.