Refactor z-score threshold implementation for mfe_window_zscore*() functions
This fixes the variant of z-score threshold MFE backtracing where only
external loop branches that satisfy the z-score threshold are taken into
account for the global solution(s). This patch also introduces a
switchable option to report structures that are actually subsumed by
another, larger structure if the z-score of the subsumed part is lower
than that of the enclosing.