Fix corner-case for pairing alignment columns with all gaps
Whether or not a pair of columns in an alignment are allowed to form
base pairs has previously been determined by the following inequality:

2 * n_c + n_g > n_seq

where n_c is the number of counter examples, i.e. the number of
sequences where i and j do not form a WC or Wobble pair, n_g is the
number of sequences where i and j would form a pair with one or two
gaps, and n_seq is the total number of sequences in the alignment.
This inequality, however, would let pass columns where not a single
canonical base pair can be formed in any sequence. By replacing
the > with a >=, we now restrict such cases, i.e. the condition to be
fulfilled is now:

2 * n_c + n_g >= n_seq