AUTOTOOLS: Always run Python3 parameter probing
This is required for any tools written in Python that we want to install
as scripts. In particular, if possible we write scripts such that they
can be loaded as part of the ViennaRNA Python bindings and install them
to the correct location. For using them as executable programs, we
resort to a wrapper shell script that simply calls the python script and
passes through any parameters. An example is the RNAconsensus.py script
found in src/Utils/