Fix circRNA predictions after change to entropic penalty and M2
This fixes the use of entropic penalty for the unpaired circular chain
in comparative structure predictions. We always take the actual length
of the individual sequences, not the length of the alignment to compute
the total penalty.

Second, this adds some missing code for M1 array filling in the case of
comparative structure predictions for circular RNA alignments.