Refactor verbose vrna_eval*() implementations
Instead of on-the-fly printing the energy contributions of individual
loop whenever verbose evaluation is done, we rather collect them and
print them in a post-processing step. This also allows for already
correcting false contributions from wrong G-Quadruplex containing loops
which previously have been printing both, the wrong contributions and
the corrections.

In addition, this would allow for passing over the individual
contributions to other functions in the future.