Be more consistent in the use of internal- vs. interior loops
This first-of all changes all unnecessary occurences of 'interior' to
'internal' in both, code and symbols. Also, this changes the energy
parameter file identifiers that have the 'interior' pre-/in-/suffix to
'internal'.

Backward compatibility is maintained for the energy parameter file
parser, but a warnign will be issued if the old identifiers are being
used.