Separate data transformation implementations from probing data stuff
This commit does a lot! It first separatess the data transformation
implementations such that they can be used to arbitrary linear data of
type double. Then, it sanitizes the interface by introducing some type
definitions and changing the order of function parameters.
This hopefully is generic enough to be reused in various future scenarios
where we want to transform linear data...